move utils to contract.ts
